qr_generator/requirements.txt
qrcode[pil]>=7.4
Pillow>=10.0.0
openpyxl>=3.1.0
requests>=2.31.0
